+++ app/build.gradle
@@ -10,14 +10,19 @@ plugins {
 android {
 
     signingConfigs {
-
+        release {
+            storeFile file('C:/Users/carlo/OneDrive/Escritorio/Key/FinderApp/FinderApp.jks')
+            storePassword 'finderApp123'
+            keyAlias 'finderApp'
+            keyPassword 'finderApp123'
+        }
     }
 
-    namespace 'com.example.finderapp'
+    namespace 'com.finder.finderapp'
     compileSdk 33
 
     defaultConfig {
-        applicationId "com.example.finderapp"
+        applicationId "com.finder.finderapp"
         minSdk 21
         targetSdk 33
         versionCode 1
@@ -37,7 +42,7 @@ android {
             minifyEnabled false
             debuggable false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-
+            signingConfig signingConfigs.release
         }
 
         debug{
@@ -68,6 +73,7 @@ dependencies {
     implementation 'com.google.android.material:material:1.9.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.core:core-ktx:1.10.1'
+    implementation 'androidx.test.ext:junit-ktx:1.1.5'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
