--- app/src/main/java/com/example/finderapp/application/AppConstants.kt
@@ -1,6 +1,0 @@
-package com.example.finderapp.application
-
-object AppConstants {
-    const val BASE_URL_YELP = "https://api.yelp.com/v3/"
-    const val BASE_URL_OPEN = "https://api.openrouteservice.org/v2/"
-}
