--- app/src/main/java/com/example/finderapp/application/BaseApplication.kt
@@ -1,6 +1,0 @@
-package com.example.finderapp.application
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-@HiltAndroidApp
-class BaseApplication: Application()
