--- app/src/main/java/com/example/finderapp/core/Qualifers.kt
@@ -1,15 +1,0 @@
-package com.example.finderapp.core
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class Yelp
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class OpenRoute
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class HttpClientOpenRoute
