--- app/src/main/java/com/example/finderapp/data/model/Permissions.kt
@@ -1,7 +1,0 @@
-package com.example.finderapp.data.model
-
-enum class Permissions {
-    REFUSED,
-    ACCEPTED,
-    PREVIOUS_VERSIONS
-}
