--- app/src/main/java/com/example/finderapp/ui/businessdetail/BusinessDetailAdapter.kt
@@ -1,42 +1,0 @@
-package com.example.finderapp.ui.businessdetail
-
-import android.annotation.SuppressLint
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import com.example.finderapp.core.loadImage
-import com.example.finderapp.core.toDateFormat
-import com.example.finderapp.data.model.Review
-import com.example.finderapp.databinding.ReviewItemBinding
-
-class BusinessDetailAdapter(private var reviewList: List<Review>): RecyclerView.Adapter<BusinessDetailAdapter.ViewHolder>() {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): BusinessDetailAdapter.ViewHolder {
-        val itemBinding = ReviewItemBinding.inflate(LayoutInflater.from(parent.context),  parent, false)
-        return  ViewHolder(itemBinding)
-    }
-
-    override fun onBindViewHolder(holder: BusinessDetailAdapter.ViewHolder, position: Int) {
-       holder.bind(reviewList[position])
-    }
-
-    override fun getItemCount(): Int {
-        return reviewList.size
-    }
-
-    @SuppressLint("NotifyDataSetChanged")
-    fun updateReviewList(reviewList: List<Review>){
-        this.reviewList = reviewList
-        notifyDataSetChanged()
-    }
-
-    inner class ViewHolder(private val itemBinding: ReviewItemBinding): RecyclerView.ViewHolder(itemBinding.root){
-        fun bind(review: Review){
-            itemBinding.imgProfile.loadImage(review.user?.image_url ?: "", true)
-            itemBinding.txtRating.text = review.rating.toString()
-            itemBinding.txtName.text = review.user?.name
-            itemBinding.txtReviewDescription.text = review.text
-            itemBinding.txtDate.text = if(!review.time_created.isNullOrEmpty()) review.time_created.toDateFormat() else ""
-        }
-    }
-}
