--- app/src/main/java/com/example/finderapp/ui/splash/SplashActivity.kt
@@ -1,16 +1,0 @@
-package com.example.finderapp.ui.splash
-
-import android.annotation.SuppressLint
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.finderapp.MainActivity
-
-@SuppressLint("CustomSplashScreen")
-class SplashActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        startActivity(Intent(this, MainActivity::class.java))
-        finish()
-    }
-}
