+++ app/src/main/java/com/finder/finderapp/MainActivity.kt
@@ -1,0 +1,33 @@
+package com.finder.finderapp
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.NavigationUI
+import com.finder.finderapp.databinding.ActivityMainBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityMainBinding
+    private lateinit var navController: NavController
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        initComponents()
+    }
+
+    private fun initComponents(){
+        val navHostFragment = supportFragmentManager.findFragmentById(R.id.fragment_container_view) as NavHostFragment
+        navController =  navHostFragment.navController
+        //set up action bar with nav controller
+        NavigationUI.setupActionBarWithNavController(this, navController)
+    }
+    override fun onSupportNavigateUp(): Boolean {
+        //For the navigation in fragments
+        return navController.navigateUp()
+    }
+}
