+++ app/src/main/java/com/finder/finderapp/application/AppConstants.kt
@@ -1,0 +1,6 @@
+package com.finder.finderapp.application
+
+object AppConstants {
+    const val BASE_URL_YELP = "https://api.yelp.com/v3/"
+    const val BASE_URL_OPEN = "https://api.openrouteservice.org/v2/"
+}
