+++ app/src/main/java/com/finder/finderapp/application/BaseApplication.kt
@@ -1,0 +1,6 @@
+package com.finder.finderapp.application
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+@HiltAndroidApp
+class BaseApplication: Application()
