+++ app/src/main/java/com/finder/finderapp/core/Qualifers.kt
@@ -1,0 +1,15 @@
+package com.finder.finderapp.core
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class Yelp
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class OpenRoute
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class HttpClientOpenRoute
