+++ app/src/main/java/com/finder/finderapp/core/ResponseResult.kt
@@ -1,0 +1,11 @@
+package com.finder.finderapp.core
+
+sealed class ResponseResult<T>(val data: T? = null, val message: String? = null){
+
+    class Success<T>(data: T): ResponseResult<T>(data)
+
+    class Error<T>(message: String?, data: T? = null): ResponseResult<T>(data, message)
+
+    class Loading<T> : ResponseResult<T>()
+
+}
