+++ app/src/main/java/com/finder/finderapp/data/model/Permissions.kt
@@ -1,0 +1,7 @@
+package com.finder.finderapp.data.model
+
+enum class Permissions {
+    REFUSED,
+    ACCEPTED,
+    PREVIOUS_VERSIONS
+}
