+++ app/src/main/java/com/finder/finderapp/repository/WebServiceOpenRoute.kt
@@ -1,0 +1,15 @@
+package com.finder.finderapp.repository
+
+import com.finder.finderapp.data.model.Route
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface WebServiceOpenRoute {
+
+    @GET("directions/driving-car")
+    suspend fun getRouteByLocation(
+        @Query("api_key") apiKey: String,
+        @Query("start", encoded = true) start: String,
+        @Query("end", encoded = true) end: String,
+    ):Route
+}
