+++ app/src/main/java/com/finder/finderapp/ui/splash/SplashActivity.kt
@@ -1,0 +1,16 @@
+package com.finder.finderapp.ui.splash
+
+import android.annotation.SuppressLint
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.finder.finderapp.MainActivity
+
+@SuppressLint("CustomSplashScreen")
+class SplashActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        startActivity(Intent(this, MainActivity::class.java))
+        finish()
+    }
+}
