+++ app/src/main/res/navigation/main_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/searchBusinessFragment"
-        android:name="com.example.finderapp.ui.searchbusiness.SearchBusinessFragment"
+        android:name="com.finder.finderapp.ui.searchbusiness.SearchBusinessFragment"
         android:label="@string/home_title"
         tools:layout="@layout/fragment_search_bussines">
         <action
@@ -20,7 +20,7 @@
     </fragment>
     <fragment
         android:id="@+id/businessDetailFragment"
-        android:name="com.example.finderapp.ui.businessdetail.BusinessDetailFragment"
+        android:name="com.finder.finderapp.ui.businessdetail.BusinessDetailFragment"
         android:label="@string/title_detail"
         tools:layout="@layout/fragment_business_detail">
         <argument
@@ -37,11 +37,11 @@
     </fragment>
     <fragment
         android:id="@+id/businessMapFragment"
-        android:name="com.example.finderapp.ui.businessmap.BusinessMapFragment"
+        android:name="com.finder.finderapp.ui.businessmap.BusinessMapFragment"
         android:label="@string/title_map"
         tools:layout="@layout/fragment_business_map">
         <argument
             android:name="Business"
-            app:argType="com.example.finderapp.data.model.Business" />
+            app:argType="com.finder.finderapp.data.model.Business" />
     </fragment>
 </navigation>
